Updated for 1.3 Removed DLLs Renamed file Added standardized build scripts
