+++ Hire/GameData/TRP-Hire/ReadMe.txt
@@ -1,0 +1,17 @@
+Version 0.6.1
+
+TRP-Hire used to be KSI-Hiring.
+TRP stands for TheReadPanda, the ID I use on twitch, twitter, youtube and the web at thereadpanda.com 
+
+The mod currently is in beta. It does not include the Mini-AVC or AVC calls.
+
+It is not up on CKAN yet and will not be until I get it out of 'beta'. But it should work. If you have issues please notify me.
+
+You can submit issues to the PHS directory for Https://github.com/TheReadPanda/PHS
+PHS stands for Panda Hiring Services.
+
+Thank you for using my mod! 
+
+Simply put the TRP-Hire folder and all the contents of it (Hire.dll, readme.txt and TRP-License.txt) into your gamedata folder for KSP. Then run KSP. If you enter the astronaut complex you should see the updated/modified UI.
+
+Happy Kerbaling!
