+++ Hire/Hire/AssemblyVersion.cs
@@ -1,0 +1,8 @@
+﻿
+ // This code was generated by a tool. Any changes made manually will be lost
+ // the next time this code is regenerated.
+ // 
+  
+ using System.Reflection;
+
+ [assembly: AssemblyVersion("0.1.0.0")]
